+++ src/util/speaker.js
@@ -13,6 +13,7 @@ exports.getUtterance = function (data, callback) {
     };
     msg.play = () => {
         // Adding custom play method to allow easier chaining with audio elements
+        speechSynthesis.cancel(); // Sometimes it gets stuck and refuses to speak any more, so cancel any existing before speaking again
         speechSynthesis.speak(msg);
     }
     return msg;
@@ -31,6 +32,7 @@ exports.getUtteranceWithVoice = function (data, voiceName, readyCallback, endCal
     };
     msg.play = () => {
         // Adding custom play method to allow easier chaining with audio elements
+        speechSynthesis.cancel(); // Sometimes it gets stuck and refuses to speak any more, so cancel any existing before speaking again
         speechSynthesis.speak(msg);
     }
     this.getVoice(voiceName, (voice) => {
@@ -66,12 +68,11 @@ exports.getVoice = function(name, callback) {
 
 exports.speak = function (data, callback) {
     const msg = this.getUtterance(data, callback);
-    speechSynthesis.cancel(); // Sometimes it gets stuck and refuses to speak any more, so cancel any existing before speaking again
-    speechSynthesis.speak(msg);
+    msg.play();
 };
 
 exports.speakWithVoice = function (data, voiceName, endCallback) {
     this.getUtteranceWithVoice(data, voiceName, (msg) => {
-        speechSynthesis.speak(msg);
+        msg.play();
     }, endCallback);
 };