+++ src/util/speaker.js
@@ -12,6 +12,7 @@ exports.getUtterance = function (data, callback) {
         }
     };
     msg.play = () => {
+        console.log(`getUtterance: speechSynthesis.speak - pending: ${speechSynthesis.pending}, speaking: ${speechSynthesis.speaking} paused: ${speechSynthesis.paused}`);
         // Adding custom play method to allow easier chaining with audio elements
         speechSynthesis.cancel(); // Sometimes it gets stuck and refuses to speak any more, so cancel any existing before speaking again
         speechSynthesis.speak(msg);
@@ -31,7 +32,7 @@ exports.getUtteranceWithVoice = function (data, voiceName, readyCallback, endCal
         }
     };
     msg.play = () => {
-        // Adding custom play method to allow easier chaining with audio elements
+        console.log(`getUtteranceWithVoice: speechSynthesis.speak - pending: ${speechSynthesis.pending}, speaking: ${speechSynthesis.speaking} paused: ${speechSynthesis.paused}`);        // Adding custom play method to allow easier chaining with audio elements
         speechSynthesis.cancel(); // Sometimes it gets stuck and refuses to speak any more, so cancel any existing before speaking again
         speechSynthesis.speak(msg);
     }