+++ CHANGELOG.md
@@ -6,6 +6,8 @@ A preview of major changes can be found in the Wiki ([Latest Changes](https://gi
 - Option to see `All` or `Unique` tournament statistics per player
 - Ability to "Remote Start" matches on a selected venue
 - Split tournaments into "Tournaments" and "Seasons"
+- Persist "Enable Voice Announcements" on spectate page
+- Added totals view per group to Tournament Overview
 
 #### Changed
 - Updated to latest version of dependencies